+++ packages/cli/CHANGELOG.md
@@ -1,10 +1,22 @@
 # @openfn/cli
 
+## 1.35.0
+
+### Minor Changes
+
+- collections: add support for --project-id
+
+### Patch Changes
+
+- Updated dependencies
+  - @openfn/lexicon@1.5.0
+
 ## 1.34.2
 
 ### Patch Changes
 
 - 3918358: Override config.endpoint with one from openfn.yaml
+
 
 ## 1.34.1